requirements.txt
asgiref==3.3.4
Django==3.2.4
django-autoslug==1.9.8
pytz==2021.1
sqlparse==0.4.1
